35 Complex Charges Detection
Started with trying to fix the charge symbol sub parts, to help with the issue of parsing consuming another charge as a subpart.
This is now fixed, actually resolving ticket #34
But, the problem now is that the different location of tincture after a symbol or after a symbol with subparts make this grammar incompatible with the multi charges grammar. That takes precedence for now (still running on the previous multi charge implementation)